requirements.txt
openai>=1.65.0
python-dotenv>=1.0.1
